+++ MetaMorpheus/EngineLayer/PrecursorSearchModes/AllowedIntervalWithNotch.cs
@@ -1,10 +1,10 @@
 ﻿namespace EngineLayer;
 
-public class AllowedIntervalWithNotch(double minimum, double maximum, int notch)
+public class AllowedIntervalWithNotch(double minimum, double maximum, double notch)
 {
     public readonly double Minimum = minimum;
     public readonly double Maximum = maximum;
-    public readonly int Notch = notch;
+    public readonly double Notch = notch;
 
     public bool Contains(double value)
     {
